feat: autolaunch & useragent settings (incomplete)
The User-Agent Settings does work, the problem is that the TizenTube module is, as the name says, is for Tizen. The User-Agent spoofs itself to be Cobalt (Google's own light HTML5 renderer) and because it has a few differences, both the back button (keyCode from Tizen is probably not there) and the TizenTube UI (need to set the nonce for styles) do not work. Once I fix this, I'll uncomment the lines and push a new update.